Stu dev 0809

+++ pay-service/pay-api/src/main/java/com/gupaoedu/pay/dto/PaymentRequest.java
@@ -1,12 +1,13 @@
 package com.gupaoedu.pay.dto;
 
 import com.gpmall.commons.result.AbstractRequest;
-import com.gpmall.pay.biz.payment.validator.validatorextend.PayChannel;
+import com.gupaoedu.pay.validatorextend.PayChannel;
 import lombok.Data;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+
 
 /**
  * 腾讯课堂搜索 咕泡学院
@@ -19,7 +20,7 @@ public class PaymentRequest extends AbstractRequest{
      * 用户id
      */
     @NotNull(message = "userId不可为空")
-    private Integer userId;
+    private Long userId;
 
     /**
      * 交易订单号, 统一生成全局唯一的订单号
